+++ workflow/scripts/arraytools.py
@@ -191,7 +191,7 @@ def relative_summit_loc(array, wsize = 50):
     peak = np.nanargmax(smoothed)
     return peak
 
-def traveling_ratio(array, wsize = 50, peak = None):
+def traveling_ratio(array, wsize = 50, peak = None, out = "ratio"):
     # if peak isn't specified then dynamically find it
     if peak is None:
         peak = relative_summit_loc(array, wsize)
@@ -206,6 +206,13 @@ def traveling_ratio(array, wsize = 50, peak = None):
         return np.nan
 
     center_avg = np.nanmean(array[(center - wsize):(center + wsize)])
+    if out == "ratio":
+        out_val = center_avg/peak_avg
+    elif out == "A":
+        out_val = peak_avg
+    elif out == "B":
+        out_val = center_avg
+    else:
+        raise ValueError("out Must be ratio, A, or B")
 
-    out = center_avg/peak_avg
-    return out    
+    return out_val