+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Spike-in based scaling using DEseq2, spike-in counts, edgeR TMM, or Bonhoure
   et al.  approaches
 - Deduplication with Picard MarkDuplicates.
+- Ability to get A and B average counts for the traveling ratio
 
 ### Changes
 - Update versions of some packages including ncbi-acc-download to get around